Add images and styling to the front page categories, streamlined product category card and product slider components.
Add images and styling to the front page category section. Streamlined product category card and product slider on front page.

The issue with the empty column on home page was caused by the react slick carousel arrows for using the carousel, disabling them in the component settings fixes the issue.