Add styling to singleview product page and fix header overflow
Add styling and placeholder images to the singleview product page and change the layout inside the single product view jsx. component to match up with the design.
Based on the design added also the tab to select whether the user wants to view the product information or availability, though the availability is currently hard coded value for all since we do not have it yet in database.

Also fixed the issue with header overflow which was again causing the blank column on the right side of the page.